+++ app/slack/templates/messages.py
@@ -3937,9 +3937,9 @@ class HumanJobQuoteMessage(SlackMessage):
                     "text": {
                         "type": "mrkdwn",
                         "text": _(
-                            "Click Accept Quote to send your translation to human "
-                            "review. Based on the AI's translation quality, a "
-                            "discount will be applied to the cost above."
+                            "Click Accept Quote to send your translation for human "
+                            "review. A discount will be applied to the quote above "
+                            "based on the quality of the AI translation."
                         ),
                     },
                 }